Login does not need to be two .tsx files.
This also fixes a state bug where 'login' would be disabled after a successful login, then a logout, and the app had not restarted.
